+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.0.2
-*INCOMPLETE*
++ Updated OpenAI client to default to local inference.
 
 ## 0.0.1
-- First pre-release.
++ First pre-release.